requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2023.10
cryptography==42.0.5
cssselect2==0.7.0
datetime-truncate==1.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
dj-easy-pdf==0.4.7
Django==5.0

django-bower==5.2.0

django-compat==1.0.15

django-import-export==3.3.7
django-jazzmin==2.6.1

django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
djangorestframework==3.14.0
et-xmlfile==1.1.0
html5lib==1.1
idna==3.6
Jinja2==3.1.3
lxml==5.1.1
MarkupPy==1.14
MarkupSafe==2.1.5
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
pillow==10.2.0
pycparser==2.21
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
pypdf==4.1.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-nvd3==0.14.2

pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
text-unidecode==1.3
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2023.4
tzlocal==5.2
Unidecode==1.3.8
uritools==4.0.2
urllib3==2.2.1
webencodings==0.5.1
xhtml2pdf==0.2.15
xlrd==2.0.1
xlwt==1.3.0